add delete mutation and type

+++ apps/server/src/resolvers/user&accounts/mutations/DeleteUserAndAccountResolver.ts
@@ -1,24 +1,28 @@
 import { Mutation, Resolver, Arg } from "type-graphql";
 import mongoose from "mongoose";
-import { typeUser } from "../../../graphqlTypes/typesUser";
-import { typeAccount } from "../../../graphqlTypes/typesAccount";
 import { verifyIfUserExist } from "../../../utils/resolvers/verifyIfUserExist";
 import { verifyIfAccountExist } from "../../../utils/resolvers/verifyIfAccountExist";
 import { deleteUser } from "../helpers/deleteUser";
 import { deleteAccount } from "../helpers/deleteAccount";
+import { typeDeleteAccount, typeDeleteUser } from "../../../graphqlTypes/typesDelete";
 
 @Resolver()
 export class DeleteUserAndAccountResolver {
   @Mutation(() => String)
   async deleteUserAndAccount(
-    @Arg("User") User: typeUser,
-    @Arg("Account") Account: typeAccount
+    @Arg("User") User: typeDeleteUser,
+    @Arg("Account") Account: typeDeleteAccount
   ): Promise<string> {
     const session = await mongoose.startSession();
     session.startTransaction();
 
-    await verifyIfUserExist(User.taxId);
-    await verifyIfAccountExist(Account.accountNumber);
+    const userExists = await verifyIfUserExist(User.taxId);
+    const accountExists = await verifyIfAccountExist(Account.accountNumber);
+    
+    if (!userExists && !accountExists) {
+      throw new Error("User and Account do not exist.");
+    }
+
 
     await deleteUser(User._id, session);
     await deleteAccount(Account._id, session);
